+++ ChangeLog.md
@@ -7,6 +7,7 @@
 + FIX: divide by 0 crash. Affects all previous releases.
   See RabbitCommon: 7ed300a206d8a36c30003b6c19ad181682d0ff5f
 + Viewer: add QCheckBox in show dialog
++ Viewer: optimize CConnecter::Name() and CConnecter::ServerName()
 
 ### v0.0.20